Should be stashed somewhere
This path ended up not working as I wanted. I may come back to it in the future. I'm not able to access the description of default values for options from here which means I would have to pass them in, which defeats the purpose of this approach. Committing as a way to save for the future. Not functional as is.